Mine Surrounding Logic (Kieran & Ahmed)
Here we have added logic for checking for tile anomolies which are the edges of the grid and then we've also began to add logic for checking all the adjacent tiles for a non-anomaly tile. All the anomolus tiles are grouped together in lists dependant on their position around the edges of the grid.